converted to 'Apps'
Each 'App' has an associated jinja2 template, and some config stored in the global site.json file.

bottle_app.py holds the route definition for eadh App, but all other code is stored in the 'theheadofabroom' package.

The Apps class is intended to be used without instantiation, all functionality is actually defined on the metaclass.

All of this is intended to make it easier to provide a pervasive look and feel.